Changed tip percentage to show integral value. Added new quotes

+++ app/src/main/java/com/merryapps/tipcalculator/ui/tip/TipUiHandler.java
@@ -69,7 +69,7 @@ public class TipUiHandler {
     }
 
     public String getTipPercentage() {
-        return this.currentTipCalculator.getTipPercentage().toString();
+        return this.currentTipCalculator.getTipPercentage().toBigInteger().toString();
     }
 
     public void setTipPercentage(String tipPercentage) {
